requirements.txt
fonttools>=4.0.0
playwright>=1.40.0
pymupdf>=1.24.0
Pillow>=10.0.0
